+++ Sources/Sentry/include/SentrySerialization.h
@@ -4,16 +4,11 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-static int const SENTRY_BAGGAGE_MAX_SIZE = 8192;
-
 @interface SentrySerialization : NSObject
 
 + (NSData *_Nullable)dataWithJSONObject:(id)jsonObject;
 
 + (NSData *_Nullable)dataWithSession:(SentrySession *)session;
-
-+ (NSDictionary<NSString *, NSString *> *)decodeBaggage:(NSString *)baggage;
-+ (NSString *)baggageEncodedDictionary:(NSDictionary *)dictionary;
 
 + (SentrySession *_Nullable)sessionWithData:(NSData *)sessionData;